+++ vlib/v/gen/native/gen.v
@@ -124,7 +124,9 @@ struct VarConfig {
 
 type Var = GlobalVar | LocalVar | ast.Ident
 
-fn (mut g Gen) get_var_from_ident(ident ast.Ident) LocalVar|GlobalVar|Register {
+type IdentVar = GlobalVar | LocalVar | Register
+
+fn (mut g Gen) get_var_from_ident(ident ast.Ident) IdentVar {
 	mut obj := ident.obj
 	if obj !in [ast.Var, ast.ConstField, ast.GlobalField, ast.AsmRegister] {
 		obj = ident.scope.find(ident.name) or { g.n_error('unknown variable $ident.name') }