+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 - DOOM is now translated/compiled and launched on CI servers. A screenshot of the running game
   is made via `vgret` and is compared to the expected result.
 - VLS performance improvements, especially on Windows.
-- Add `v ls` tool for installing, for updating, and for launching VLS (V Language Server)
+- `v ls` tool for installing, for updating, and for launching VLS (V Language Server).
 - Support `assert condition, extra_message`, where the `extra_message` will be evaluated and shown if the assertion fails.
+- Anonymous sumtypes have been removed (deprecated for now) due to complicating the language and the compiler too much.
 
 ## V 0.3
 *30 Jun 2022*